Remove jskit session workflow
