+++ packages/agent-docs/test/appAgentTemplate.test.js
@@ -7,17 +7,14 @@ import { fileURLToPath } from "node:url";
 const packageRoot = path.resolve(path.dirname(fileURLToPath(import.meta.url)), "..");
 const templatePath = path.join(packageRoot, "templates/app/AGENTS.md");
 
-test("app agent template points to jskit session instead of old prose workflows", async () => {
+test("app agent template points to distributed docs without removed workflow commands", async () => {
   const body = await readFile(templatePath, "utf8");
 
-  assert.match(body, /jskit session create/);
-  assert.match(body, /jskit session <session_id> step/);
-  assert.match(body, /--json/);
-  assert.match(body, /Do not invent a parallel manual issue workflow/);
+  assert.match(body, /agent-docs\/guide\/agent\/index\.md/);
+  assert.match(body, /agent-docs\/patterns\/INDEX\.md/);
 
   assert.doesNotMatch(body, /workflow\/app-state\.md/);
   assert.doesNotMatch(body, /workflow\/feature-delivery\.md/);
   assert.doesNotMatch(body, /Mandatory Start Gate/);
   assert.doesNotMatch(body, /Mandatory Done Gate/);
-  assert.doesNotMatch(body, /WORKBOARD/);
 });
