+++ packages/agent-docs/test/crudGuidance.test.js
@@ -6,13 +6,11 @@ import { fileURLToPath } from "node:url";
 
 const packageRoot = path.resolve(path.dirname(fileURLToPath(import.meta.url)), "..");
 
-test("distributed agent docs no longer expose workflow files as active guidance", async () => {
+test("distributed agent docs no longer expose removed workflow files or commands", async () => {
   const packageJson = JSON.parse(await readFile(path.join(packageRoot, "package.json"), "utf8"));
-  const distributedGuide = await readFile(path.join(packageRoot, "DISTR_AGENT.md"), "utf8");
+  const agentGuide = await readFile(path.join(packageRoot, "guide/agent/index.md"), "utf8");
 
   assert.doesNotMatch(JSON.stringify(packageJson.files), /workflow/);
-  assert.match(distributedGuide, /jskit session create/);
-  assert.match(distributedGuide, /Do not use the old prose workflow model/);
-  assert.doesNotMatch(distributedGuide, /workflow\/scoping\.md/);
-  assert.doesNotMatch(distributedGuide, /workflow\/review\.md/);
+  assert.doesNotMatch(agentGuide, /workflow\/scoping\.md/);
+  assert.doesNotMatch(agentGuide, /workflow\/review\.md/);
 });
