+++ tooling/jskit-cli/src/server/core/argParser.js
@@ -23,8 +23,6 @@ function parseArgs(argv, { createCliError } = {}) {
         verbose: false,
         json: false,
         all: false,
-        abandoned: false,
-        completed: false,
         concrete: false,
         help: true,
         inlineOptions: {}
@@ -52,8 +50,6 @@ function parseArgs(argv, { createCliError } = {}) {
     verbose: false,
     json: false,
     all: false,
-    abandoned: false,
-    completed: false,
     concrete: false,
     help: false,
     inlineOptions: {}
@@ -111,14 +107,6 @@ function parseArgs(argv, { createCliError } = {}) {
     }
     if (token === "--all") {
       options.all = true;
-      continue;
-    }
-    if (token === "--abandoned") {
-      options.abandoned = true;
-      continue;
-    }
-    if (token === "--completed") {
-      options.completed = true;
       continue;
     }
     if (token === "--concrete") {
