+++ tooling/jskit-cli/src/server/index.js
@@ -1,3 +1,2 @@
 export { runCli } from "./core/createCliRunner.js";
-export * from "./sessionRuntime.js";
 export * from "./appBlueprint.js";
