--- tooling/jskit-cli/src/server/sessionRuntime/prompts/final_comment.md
@@ -1,10 +1,0 @@
-Codex thread id: {{codex_thread_id}}
-Issue session status: finished
-PR: {{pr_url}}
-PR outcome: {{pr_outcome}}
-Outcome reason: {{pr_outcome_reason}}
-Local transcript: {{transcript_log}}
-
-Summary:
-
-Session {{session_id}} finished with PR outcome `{{pr_outcome}}`.
