--- tooling/jskit-cli/src/server/sessionRuntime/prompts/review_changes_accepted_resolve.md
@@ -1,12 +1,0 @@
-Resolve the important findings from the latest deslop review in this conversation.
-
-Fix only those findings. Keep scope tight. Do not create commits, branches, issues, pull requests, merges, or worktree cleanup.
-
-Use existing JSKIT helpers, generators, package seams, and app-local helpers where they fit. Do not introduce helper churn or unrelated refactors while resolving these findings.
-
-When finished:
-
-- Summarize the exact findings fixed.
-- List changed files.
-- List checks run, or say why checks were not run.
-- Wait for the next review prompt.
